회원가입 프론트엔드 구현중

+++ backend/src/main/java/com/byul/wheresmymoney/backend/User/Entity/UserEntity.java
@@ -1,7 +1,18 @@
 package com.byul.wheresmymoney.backend.User.Entity;
 
-import jakarta.persistence.*;
-import lombok.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 
 @Entity
@@ -15,7 +26,7 @@ public class UserEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "userinfo_idx")
-    private int userinfo_idx;
+    private long userinfo_idx;
 
     // ID는 15자 제한.
     @Column(unique = true, nullable = true, length = 15)
